refactor(core/oauth): rename Token to OAuthToken
This commit renames the `Token` data class to `OAuthToken` within the `core.oauth.model` package.

This change provides better clarity regarding the purpose of the class, specifically indicating that it represents an OAuth token.

The following files and their usages of `Token` have been updated to reflect the new name `OAuthToken`:
- `core/oauth/src/main/kotlin/com/music/dzr/core/oauth/model/OAuthToken.kt` (renamed from `Token.kt`)
- `core/oauth/src/main/kotlin/com/music/dzr/core/oauth/repository/TokenRepository.kt`
- `core/network/src/main/kotlin/com/music/dzr/core/network/model/auth/Token.kt`
- `core/testing/src/main/kotlin/com/music/dzr/core/testing/repository/FakeTokenRepository.kt`